[TASK]: split js code into modules
